Cleanup device page menus
