+++ frontend/src/components/DataDisplay/DataDisplay.tsx
@@ -9,7 +9,7 @@ export const DataDisplay: React.FC<{ data: IDataDisplay[] }> = ({ data }) => {
   const css = useStyles()
 
   return (
-    <List>
+    <List className={css.list}>
       {data.map(
         item =>
           item.value != null && (
@@ -31,6 +31,7 @@ export const DataDisplay: React.FC<{ data: IDataDisplay[] }> = ({ data }) => {
 }
 
 const useStyles = makeStyles({
+  list: { width: '100%' },
   item: {
     padding: `4px 0`,
     fontSize: fontSizes.sm,
