+++ frontend/src/components/ListItemLocation/ListItemLocation.tsx
@@ -1,12 +1,12 @@
 import React from 'react'
-import { useHistory } from 'react-router-dom'
-import { ListItem, ListItemIcon, ListItemText } from '@material-ui/core'
+import { useHistory, useLocation } from 'react-router-dom'
+import { makeStyles, ListItem, ListItemIcon, ListItemText } from '@material-ui/core'
 import { Icon } from '../Icon'
-import { Color } from '../../styling'
+import { colors, Color } from '../../styling'
 
 export type Props = {
   pathname: string
-  title?: string
+  title?: React.ReactElement | string
   subtitle?: string
   icon?: string
   iconColor?: Color
@@ -25,10 +25,22 @@ export const ListItemLocation: React.FC<Props> = ({
   children,
   ...props
 }) => {
+  const css = useStyles()
   const history = useHistory()
+  const location = useLocation()
+  const active = location.pathname === pathname
+
   const onClick = () => !disabled && history.push(pathname)
   return (
-    <ListItem {...props} button onClick={onClick} disabled={disabled} style={{ opacity: 1 }}>
+    <ListItem
+      {...props}
+      button
+      selected={active}
+      onClick={onClick}
+      disabled={disabled}
+      style={{ opacity: 1 }}
+      classes={{ selected: css.selected }}
+    >
       {icon && (
         <ListItemIcon>
           <Icon name={icon} size="md" color={iconColor} fixedWidth />
@@ -39,3 +51,10 @@ export const ListItemLocation: React.FC<Props> = ({
     </ListItem>
   )
 }
+
+const useStyles = makeStyles({
+  selected: {
+    '&.Mui-selected': { backgroundColor: colors.primaryHighlight },
+    '&.Mui-selected:hover': { backgroundColor: colors.primaryLighter },
+  },
+})
