+++ frontend/src/components/Panel.tsx
@@ -6,15 +6,15 @@ import { Header } from './Header'
 
 type Props = {
   primary?: boolean
-  secondary?: boolean
+  resize?: boolean
 }
 
-export const Panel: React.FC<Props> = ({ primary, secondary, children }) => {
+export const Panel: React.FC<Props> = ({ primary, resize, children }) => {
   const [width, setWidth] = React.useState<number>(400)
   const [grab, setGrab] = React.useState<boolean>(false)
-  const css = useStyles()
+  const css = useStyles(primary)()
 
-  const onMove = event => grab && setWidth(width - event.movementX * 2)
+  const onMove = event => grab && setWidth(width + event.movementX * 2)
   const onDown = () => setGrab(true)
   const onUp = () => setGrab(false)
 
@@ -29,46 +29,48 @@ export const Panel: React.FC<Props> = ({ primary, secondary, children }) => {
 
   return (
     <>
-      {secondary && (
-        <div className={css.handle} onMouseDown={onDown}>
-          <div className={grab ? 'active' : undefined} />
-        </div>
-      )}
       <div className={css.panel} style={{ width }}>
         {primary && <Header />}
         {children}
       </div>
+      {resize && (
+        <div className={css.handle} onMouseDown={onDown}>
+          <div className={grab ? 'active' : undefined} />
+        </div>
+      )}
     </>
   )
 }
 
-const useStyles = makeStyles({
-  panel: {
-    flexGrow: 1,
-    height: '100%',
-    display: 'flex',
-    flexDirection: 'column',
-  },
-  handle: {
-    height: '100%',
-    padding: `${0} ${spacing.sm}px`,
-    '-webkit-app-region': 'no-drag',
-    '&:hover': {
-      cursor: 'col-resize',
-    },
-    '& > div': {
-      width: 1,
-      marginLeft: 1,
-      marginRight: 1,
+const useStyles = primary =>
+  makeStyles({
+    panel: {
+      flexGrow: 1,
       height: '100%',
-      backgroundColor: colors.grayLighter,
-      transition: 'background-color 100ms 200ms, width 100ms 200ms, margin 100ms 200ms',
+      display: 'flex',
+      flexDirection: 'column',
+      marginTop: primary ? undefined : spacing.xl,
     },
-    '&:hover > div, & .active': {
-      width: 3,
-      marginLeft: 0,
-      marginRight: 0,
-      backgroundColor: colors.primary,
+    handle: {
+      height: '100%',
+      padding: `${0} ${spacing.xs}px`,
+      '-webkit-app-region': 'no-drag',
+      '&:hover': {
+        cursor: 'col-resize',
+      },
+      '& > div': {
+        width: 1,
+        marginLeft: 1,
+        marginRight: 1,
+        height: '100%',
+        backgroundColor: colors.grayLighter,
+        transition: 'background-color 100ms 200ms, width 100ms 200ms, margin 100ms 200ms',
+      },
+      '&:hover > div, & .active': {
+        width: 3,
+        marginLeft: 0,
+        marginRight: 0,
+        backgroundColor: colors.primary,
+      },
     },
-  },
-})
+  })
