--- frontend/src/components/Router/index.ts
@@ -1,1 +1,0 @@
-export { Router } from './Router'
