+++ frontend/src/components/ServiceHeaderMenu.tsx
@@ -1,59 +1,37 @@
-import React, { useState, useEffect } from 'react'
-import { useSelector, useDispatch } from 'react-redux'
-import { useParams, useLocation } from 'react-router-dom'
+import React, { useState } from 'react'
+import { useSelector } from 'react-redux'
+import { useParams } from 'react-router-dom'
+import { OutOfBand } from './OutOfBand'
 import { makeStyles } from '@material-ui/core/styles'
-import { PortSetting } from './PortSetting'
-import { NameSetting } from './NameSetting'
 import { ServiceName } from './ServiceName'
-import { Breadcrumbs } from './Breadcrumbs'
-import { HostSetting } from './HostSetting'
-import { ProxySetting } from './ProxySetting'
-import { selectService } from '../models/devices'
 import { LicensingNotice } from './LicensingNotice'
 import { ListItemLocation } from './ListItemLocation'
-import { ServiceConnected } from './ServiceConnected'
-import { AutoStartSetting } from './AutoStartSetting'
-import { CustomAttributeSettings } from './CustomAttributeSettings'
-import { ApplicationState, Dispatch } from '../store'
+import { ApplicationState } from '../store'
 import { Typography, Divider, List } from '@material-ui/core'
 import { ConnectionErrorMessage } from './ConnectionErrorMessage'
-import { InlineTemplateSetting } from './InlineTemplateSetting'
-import { ConnectionLogSetting } from './ConnectionLogSetting'
+import { UnregisterServiceButton } from '../buttons/UnregisterServiceButton'
+import { DeleteServiceButton } from '../buttons/DeleteServiceButton'
 import { ConnectionStateIcon } from './ConnectionStateIcon'
 import { UnauthorizedPage } from '../pages/UnauthorizedPage'
-import { LanShareSelect } from './LanShareSelect'
-import { LoadingMessage } from './LoadingMessage'
 import { AddUserButton } from '../buttons/AddUserButton'
-import { ConnectButton } from '../buttons/ConnectButton'
-import { LaunchButton } from '../buttons/LaunchButton'
-import { ForgetButton } from '../buttons/ForgetButton'
 import { UsersSelect } from './UsersSelect'
 import { ErrorButton } from '../buttons/ErrorButton'
-import { EditButton } from '../buttons/EditButton'
-import { CopyButton } from '../buttons/CopyButton'
 import { Container } from './Container'
-import { Columns } from './Columns'
-import { spacing } from '../styling'
-import analyticsHelper from '../helpers/analyticsHelper'
 
-export const ServiceHeaderMenu: React.FC = () => {
+export const ServiceHeaderMenu: React.FC<{ device?: IDevice; service?: IService; target?: ITarget }> = ({
+  device,
+  service,
+  target,
+  children,
+}) => {
   const css = useStyles()
-  const location = useLocation()
-  const { deviceID, serviceID = '' } = useParams<{ deviceID: string; serviceID: string }>()
+  const { serviceID = '' } = useParams<{ deviceID: string; serviceID: string }>()
   const [showError, setShowError] = useState<boolean>(true)
-  const { devices } = useDispatch<Dispatch>()
-  const { connection, service, device, thisDevice, fetching, access } = useSelector((state: ApplicationState) => {
-    const connection = state.backend.connections.find(c => c.id === serviceID)
-    const [service, device] = selectService(state, serviceID)
-    return {
-      service,
-      device,
-      connection,
-      thisDevice: state.backend.device?.uid === device?.id,
-      fetching: state.devices.fetching,
-      access: state.accounts.access,
-    }
-  })
+  const { connection, thisDevice, access } = useSelector((state: ApplicationState) => ({
+    connection: state.backend.connections.find(c => c.id === serviceID),
+    thisDevice: state.backend.device?.uid === device?.id,
+    access: state.accounts.access,
+  }))
 
   // useEffect(() => {
   //   analyticsHelper.page('ServicePage')
@@ -68,14 +46,35 @@ export const ServiceHeaderMenu: React.FC = () => {
     <Container
       header={
         <>
+          <OutOfBand />
           <Typography variant="h1">
-            <ConnectionStateIcon connection={connection} service={service} thisDevice={thisDevice} size="lg" />
-            <ServiceName connection={connection} service={service} inline />
+            {/* <ConnectionStateIcon connection={connection} service={service} thisDevice={thisDevice} size="lg" /> */}
+            <ServiceName connection={connection} service={service} />
             <ErrorButton connection={connection} onClick={() => setShowError(!showError)} visible={showError} />
             <AddUserButton device={device} />
-            <LaunchButton connection={connection} service={service} />
-            <CopyButton connection={connection} service={service} />
+            {thisDevice ? (
+              <UnregisterServiceButton target={target} />
+            ) : (
+              <DeleteServiceButton device={device} service={service} />
+            )}
           </Typography>
+          <List>
+            {!device.shared && (
+              <ListItemLocation
+                title="Edit Service"
+                icon="pen"
+                pathname={`/devices/${device.id}/${serviceID}/edit`}
+                dense
+              />
+            )}
+            <UsersSelect service={service} device={device} access={access} />
+            <ListItemLocation
+              title="Service Details"
+              icon="info-circle"
+              pathname={`/devices/${device.id}/${serviceID}/details`}
+              dense
+            />
+          </List>
           <List className={css.errorMessage}>
             <ConnectionErrorMessage connection={connection} service={service} visible={showError} />
           </List>
@@ -83,18 +82,7 @@ export const ServiceHeaderMenu: React.FC = () => {
         </>
       }
     >
-      <List>
-        {!device.shared && (
-          <ListItemLocation title="Edit Service" icon="pen" pathname={`/devices/${deviceID}/${serviceID}/edit`} dense />
-        )}
-        <UsersSelect service={service} device={device} access={access} />
-        <ListItemLocation
-          title="Service Details"
-          icon="info-circle"
-          pathname={`/devices/${deviceID}/${serviceID}/details`}
-          dense
-        />
-      </List>
+      {children}
     </Container>
   )
 }
