--- frontend/src/pages/DeviceEditPage/index.ts
@@ -1,1 +1,0 @@
-export { DeviceEditPage } from './DeviceEditPage'
