+++ frontend/src/pages/ServiceEditPage/ServiceEditPage.tsx
@@ -1,15 +1,12 @@
 import React, { useEffect } from 'react'
 import { useHistory, useLocation } from 'react-router-dom'
-import { selectService } from '../../models/devices'
+import { selectById } from '../../models/devices'
 import { useSelector, useDispatch } from 'react-redux'
 import { ApplicationState, Dispatch } from '../../store'
-import { UnregisterServiceButton } from '../../buttons/UnregisterServiceButton'
-import { DeleteServiceButton } from '../../buttons/DeleteServiceButton'
 import { REGEX_LAST_PATH } from '../../shared/constants'
 import { Typography } from '@material-ui/core'
 import { useParams } from 'react-router-dom'
-import { Container } from '../../components/Container'
-import { OutOfBand } from '../../components/OutOfBand'
+import { ServiceHeaderMenu } from '../../components/ServiceHeaderMenu'
 import { Breadcrumbs } from '../../components/Breadcrumbs'
 import { ServiceForm } from '../../components/ServiceForm'
 import { getLinks } from '../../helpers/routeHelper'
@@ -20,18 +17,15 @@ import analyticsHelper from '../../helpers/analyticsHelper'
 type Props = {
   targets: ITarget[]
   targetDevice: ITargetDevice
+  device?: IDevice
 }
-export const ServiceEditPage: React.FC<Props> = ({ targets, targetDevice }) => {
+export const ServiceEditPage: React.FC<Props> = ({ targets, targetDevice, device }) => {
   const { devices, backend, applicationTypes } = useDispatch<Dispatch>()
   const { serviceID = '', deviceID } = useParams<{ serviceID: string; deviceID: string }>()
-  const { device, service, links } = useSelector((state: ApplicationState) => {
-    const [service, device] = selectService(state, serviceID)
-    return {
-      device,
-      service,
-      links: getLinks(state, deviceID),
-    }
-  })
+  const { service, links } = useSelector((state: ApplicationState) => ({
+    service: device?.services.find(s => s.id === serviceID),
+    links: getLinks(state, deviceID),
+  }))
   const target = targets?.find(t => t.uid === serviceID)
   const thisDevice = service?.deviceID === targetDevice.uid
   const location = useLocation()
@@ -47,26 +41,10 @@ export const ServiceEditPage: React.FC<Props> = ({ targets, targetDevice }) => {
     return null
   }
 
-  console.log('SERVICE ID', serviceID)
   const exit = () => history.push(location.pathname.replace(REGEX_LAST_PATH, ''))
 
   return (
-    <Container
-      header={
-        <>
-          <OutOfBand />
-          <Typography variant="h1">
-            <Icon name="pen" size="lg" type="light" color="grayDarker" fixedWidth />
-            <Title inline>Edit service</Title>
-            {thisDevice ? (
-              <UnregisterServiceButton target={target} />
-            ) : (
-              <DeleteServiceButton device={device} service={service} />
-            )}
-          </Typography>
-        </>
-      }
-    >
+    <ServiceHeaderMenu device={device} service={service} target={target}>
       <ServiceForm
         service={service}
         target={target}
@@ -90,6 +68,6 @@ export const ServiceEditPage: React.FC<Props> = ({ targets, targetDevice }) => {
           exit()
         }}
       />
-    </Container>
+    </ServiceHeaderMenu>
   )
 }
