--- frontend/src/pages/ServicesPageTO-DELLETE/index.ts
@@ -1,1 +1,0 @@
-export { ServicesPage } from './ServicesPage'
