+++ frontend/src/pages/UsersPageDevice/UsersPageDevice.tsx
@@ -1,20 +1,13 @@
 import React, { useEffect } from 'react'
 import { ApplicationState } from '../../store'
-import { SharedUsersHeader } from '../../components/SharedUsersHeader'
 import { selectSessionUsers } from '../../models/sessions'
 import { SharedUsersList } from '../../components/SharedUsersList'
+import { DeviceHeaderMenu } from '../../components/DeviceHeaderMenu'
 import { useSelector } from 'react-redux'
-import { getDevices } from '../../models/accounts'
-import { useParams } from 'react-router-dom'
-import { Container } from '../../components/Container'
 import analyticsHelper from '../../helpers/analyticsHelper'
 
-export const UsersPageDevice: React.FC = () => {
-  const { deviceID = '' } = useParams<{ deviceID: string }>()
-  const { device, connected } = useSelector((state: ApplicationState) => ({
-    device: getDevices(state).find((d: IDevice) => d.id === deviceID),
-    connected: selectSessionUsers(state, deviceID),
-  }))
+export const UsersPageDevice: React.FC<{ device?: IDevice }> = ({ device }) => {
+  const connected = useSelector((state: ApplicationState) => selectSessionUsers(state, device?.id))
   const users = device?.access
 
   useEffect(() => {
@@ -22,8 +15,8 @@ export const UsersPageDevice: React.FC = () => {
   }, [])
 
   return (
-    <Container header={<SharedUsersHeader device={device} />}>
+    <DeviceHeaderMenu device={device}>
       <SharedUsersList device={device} users={users} connected={connected} />
-    </Container>
+    </DeviceHeaderMenu>
   )
 }
