+++ frontend/src/pages/UsersPageService/UsersPageService.tsx
@@ -1,26 +1,18 @@
 import React, { useEffect } from 'react'
 import { useParams } from 'react-router-dom'
-import { Container } from '../../components/Container'
-import { getDevices } from '../../models/accounts'
-import { findService } from '../../models/devices'
 import { useSelector } from 'react-redux'
-import { attributeName } from '../../shared/nameHelper'
 import { SharedUsersList } from '../../components/SharedUsersList'
 import { ApplicationState } from '../../store'
-import { SharedUsersHeader } from '../../components/SharedUsersHeader'
+import { ServiceHeaderMenu } from '../../components/ServiceHeaderMenu'
 import { selectSessionUsers } from '../../models/sessions'
 import analyticsHelper from '../../helpers/analyticsHelper'
 
-export const UsersPageService: React.FC = () => {
+export const UsersPageService: React.FC<{ device?: IDevice }> = ({ device }) => {
   const { serviceID = '' } = useParams<{ serviceID: string }>()
-  const { service, device, connected } = useSelector((state: ApplicationState) => {
-    const [service, device] = findService(getDevices(state), serviceID)
-    return {
-      connected: selectSessionUsers(state, serviceID),
-      service,
-      device,
-    }
-  })
+  const { service, connected } = useSelector((state: ApplicationState) => ({
+    connected: selectSessionUsers(state, serviceID),
+    service: device?.services.find(s => s.id === serviceID),
+  }))
   const users = service?.access
 
   useEffect(() => {
@@ -28,8 +20,8 @@ export const UsersPageService: React.FC = () => {
   }, [])
 
   return (
-    <Container header={<SharedUsersHeader device={device} title={`${attributeName(service)} users`} />}>
+    <ServiceHeaderMenu device={device} service={service}>
       <SharedUsersList users={users} connected={connected} />
-    </Container>
+    </ServiceHeaderMenu>
   )
 }
