+++ frontend/src/styling/index.ts
@@ -44,7 +44,7 @@ export type Color =
 export const colors: { [key in Color]: string } = {
   primary: '#0096e7',
   primaryLight: '#9ed3f0',
-  primaryLighter: '#e7f2f9',
+  primaryLighter: '#d7effc',
   primaryHighlight: '#edf8ff',
   secondary: '#75bd00',
   successLight: '#a4db4c',
